Refined UI components and improved media player precision.
- Improved video player progress tracking by increasing update frequency and switching to double-precision percentage calculations.
- Added a play icon overlay to video thumbnails on the main screen for better visual identification.
- Switched media card dimensions to use an aspect ratio modifier instead of fixed height and width.
- Removed redundant height constraints and hardcoded sizes in `QuickSaveSettingsScreen`, `SavedStatusScreen`, and `OnBoardingDialog`.
- Cleaned up unused variables, log statements, and flags in `QuickSaveSettingsViewModel` and `StatusesManagerUseCase`.

+++ app/src/main/java/com/voiddevelopers/wastatussaver/presentation/ui/player/ui/videoAudioPlayerRoot/AudioVideoPlayerViewModel.kt
@@ -230,8 +230,8 @@ class AudioVideoPlayerViewModel @Inject constructor(
         val position = player.currentPosition
 
         val percent = if (duration > 0) {
-            (position * 100 / duration)
-        } else 0L
+            (position * 100.0 / duration)
+        } else 0.0
 
         updateUiState {
             copy(
@@ -281,7 +281,7 @@ data class AudioVideoPlayerUiState(
     val mediaInfo: MediaInfo = emptyMediaInfo,
     var isPlaying: Boolean = false,
     val mediaEnded: Boolean = false,
-    val progressPercent: Long = 0L,
+    val progressPercent: Double = 0.0,
     val durationMs: Long = 0L,
     val showControls: Boolean = true,
     val downloadState: DownloadState = DownloadState.NOT_DOWNLOADED,
